requirements.txt
accelerate==0.22.0
datasets==2.14.3
deepspeed==0.10.2
Flask==2.3.2
ipdb==0.13.13
nltk==3.8.1
numpy==1.24.3
peft==0.5.0
PyYAML==6.0.1
torch==1.13.0
tqdm==4.65.0
transformers==4.32.0
pydantic==1.8.2
